Created basic logic flow (my-channel)
