+++ src/app/channel/create/page.tsx
@@ -4,18 +4,14 @@ import CreateChannel from "@/components/my-channel/create-channel";
 import { useUser } from "@/lib/context/user-context";
 
 export default function CreateChannelPage() {
-  const { user, isLoading } = useUser();
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
+  const { user } = useUser();
 
   if (!user || user.role !== "editor") {
     return <div>You are not authorized to create a channel.</div>;
   }
 
   return (
-    <div>
+    <div className="p-4">
       <CreateChannel />
     </div>
   );
