+++ src/app/my-channel/full/page.tsx
@@ -3,19 +3,19 @@
 import { useUser } from "@/lib/context/user-context";
 
 export default function MyChannelFull() {
-  const { user, isLoading } = useUser();
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
+  const { user } = useUser();
 
   if (!user?.subscriptionId) {
-    return <div>You must have a subscription to view the full content.</div>;
+    return (
+      <div className="p-4">
+        <p>You must have a subscription to view the full content.</p>
+      </div>
+    );
   }
 
   return (
-    <div>
-      <h2>My Channel (Full Access)</h2>
+    <div className="p-4">
+      <h2 className="text-2xl font-bold mb-4">My Channel (Full Access)</h2>
       <p>Full article content goes here.</p>
     </div>
   );
